发布 1.1.8-beta NeoForge / Release 1.1.8-beta for NeoForge
完成配置归属与服务端同步、建造取料一致性、工作流原因、存储分页、蓝图预览缓存和范围选择一致性；同步测试与技术附录。

Completes server-owned configuration sync, consistent construction material sourcing, workflow reasons, storage paging, blueprint preview caching, and selection-limit parity, with matching tests and technical appendix updates.

+++ src/main/java/com/rtsbuilding/rtsbuilding/client/screen/quickbuild/QuickBuildConvenienceController.java
@@ -109,7 +109,7 @@ final class QuickBuildConvenienceController {
         QuickBuildUiConvenienceSettings settings = preferences.convenienceSettings();
         return switch (preferences.convenienceTool()) {
             case REPEAT_BOX -> settings.sizeX() + "×" + settings.sizeY() + "×" + settings.sizeZ();
-            case CHUNK_QUARRY -> "16×" + (settings.chunkUp() + settings.chunkDown() + 1) + "×16";
+            case CHUNK_QUARRY -> "16×" + ((long) settings.chunkUp() + settings.chunkDown() + 1L) + "×16";
             case TREE_FELL -> "≤ " + settings.treeMaxBlocks();
         };
     }
